Fix type propagation in with-final! with-open!
Previously a type hint on the initialization would not propagate to
the eventual call site for the :always or :error handler which made it
impossible to avoid reflection warnings.

Change the expansion to drop the initialization value's redirection
through an anonymous function argument. This allows any type hint in
the original form to take effect.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>